optimize dygraph performance with move runtime import to begining (#37759)
* optimize dygraph probl

* refine code

* fix convert dtype error

* fix import datafeeder error
